+++ orbit/pkg/table/containerd/mounts_linux.go
@@ -1,7 +1,7 @@
 //go:build linux
 // +build linux
 
-package containerd_mounts
+package containerd
 
 import (
 	"context"
@@ -13,8 +13,8 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-// Columns is the schema of the table.
-func Columns() []table.ColumnDefinition {
+// MountsColumns is the schema of the containerd_mounts table.
+func MountsColumns() []table.ColumnDefinition {
 	return []table.ColumnDefinition{
 		table.TextColumn("namespace"),
 		table.TextColumn("container_id"),
@@ -25,9 +25,9 @@ func Columns() []table.ColumnDefinition {
 	}
 }
 
-// Generate is called to return the results for the table at query time.
+// GenerateMounts is called to return the results for the containerd_mounts table at query time.
 // Constraints for generating can be retrieved from the queryContext.
-func Generate(ctx context.Context, queryContext table.QueryContext) ([]map[string]string, error) {
+func GenerateMounts(ctx context.Context, queryContext table.QueryContext) ([]map[string]string, error) {
 	client, err := containerd.New("/run/containerd/containerd.sock")
 	if err != nil {
 		return nil, fmt.Errorf("Failed to connect to containerd: %v", err)
