+++ orbit/pkg/table/extension_linux.go
@@ -5,8 +5,7 @@ package table
 import (
 	"context"
 
-	"github.com/fleetdm/fleet/v4/orbit/pkg/table/containerd_containers"
-	"github.com/fleetdm/fleet/v4/orbit/pkg/table/containerd_mounts"
+	"github.com/fleetdm/fleet/v4/orbit/pkg/table/containerd"
 	"github.com/fleetdm/fleet/v4/orbit/pkg/table/crowdstrike/falcon_kernel_check"
 	"github.com/fleetdm/fleet/v4/orbit/pkg/table/crowdstrike/falconctl"
 	"github.com/fleetdm/fleet/v4/orbit/pkg/table/cryptsetup"
@@ -27,8 +26,8 @@ func PlatformTables(opts PluginOpts) ([]osquery.OsqueryPlugin, error) {
 		falcon_kernel_check.TablePlugin(log.Logger),   // table name is "falcon_kernel_check"
 		dataflattentable.TablePluginExec(log.Logger, "nftables", dataflattentable.JsonType, []string{"nft", "-jat", "list", "ruleset"}, dataflattentable.WithBinDirs("/usr/bin", "/usr/sbin")), // -j (json) -a (show object handles) -t (terse, omit set contents)
 		table.NewPlugin("dconf_read", dconf_read.Columns(), dconf_read.Generate),
-		table.NewPlugin("containerd_containers", containerd_containers.Columns(), containerd_containers.Generate),
-		table.NewPlugin("containerd_mounts", containerd_mounts.Columns(), containerd_mounts.Generate),
+		table.NewPlugin("containerd_containers", containerd.ContainersColumns(), containerd.GenerateContainers),
+		table.NewPlugin("containerd_mounts", containerd.MountsColumns(), containerd.GenerateMounts),
 		table.NewPlugin(fleetd_pacman_packages.TableName, fleetd_pacman_packages.Columns(), fleetd_pacman_packages.Generate),
 		table.NewPlugin("crowdstrike_falcon", crowdstrike_falcon.CrowdstrikeFalconColumns(),
 			func(ctx context.Context, queryContext table.QueryContext) ([]map[string]string, error) {
